+++ tests/phpunit/Support/CiviApi4Stubs.php
@@ -30,11 +30,25 @@ namespace Civi\Api4\Generic {
 namespace {
   if (!class_exists('CRM_Utils_Hook')) {
     class CRM_Utils_Hook {
+      /** @var array<string, callable> */
+      private static array $callbacks = [];
+
       public static function singleton(): self {
         return new self();
       }
 
+      public static function setCallback(string $hook, callable $callback): void {
+        self::$callbacks[$hook] = $callback;
+      }
+
+      public static function resetCallbacks(): void {
+        self::$callbacks = [];
+      }
+
       public function invoke(array $names, &$arg1, &$arg2, &$arg3, &$arg4, &$arg5, &$arg6, string $hook): void {
+        if (isset(self::$callbacks[$hook])) {
+          self::$callbacks[$hook]($arg1, $arg2, $arg3, $arg4, $arg5, $arg6);
+        }
       }
     }
   }